Install vinValidator create middleware

+++ data/migrations/01-make_cars_table.js
@@ -5,7 +5,7 @@ exports.up = function(knex) {
         tbl.string('vin').unique().notNullable();
         tbl.string('make').notNullable();
         tbl.string('model').notNullable();
-        tbl.numeric('mileage').notNullable();
+        tbl.integer('mileage').notNullable();
         tbl.string('title');
         tbl.string('transmission');
     })
